Nebra Helium Bluetooth issue
Hi everyone. New to the Helium network and to some capacity a tech noob as well, so bare with me..

I was mining with my Nebra purchased new off craigslist for about 6 days last week. It went off line and said "needs attention" I have tried to get in to diagnostics through bluetooth (and through the IP in case that was your next suggestion, and currently the ip address in my browser doesn't even show up) I had some difficulty already with the bluetooth pairing. The bluetooth button specifically will not activate bluetooth. I hold it for 5 seconds - 10 seconds and nothing. Up until now I was only able to get bluetooth turned on after restarting the hotspot and was able to find it to pair ONCE per restart on bluetooth. If I tried searching for my device after leaving the diagnostics page it would not pair with bluetooth again.

Troubleshooting steps I have taken;

I have deleted the app, shut off my phones bluetooth, closed the app, unplugged the hot spot 10-15 times in a row to prompt bluetooth to engage, and nothing will activate bluetooth anymore. Pulled the cover off and confirmed everything to my knowledge was pushed in and seated properly.

Also of note, I plugged an ethernet cable in to the hotspot and at first the ethernet port lit up and showed connectivity, but as of yesterday no longer even blinks or shows connected.

I found some mentions of Emmc card/SD card issues, do any of you tech savvy people think that could be my bluetooth pairing issue, and if so what in layman's terms do I need to do to fix it.

Thanks for any advice or support. love that at least i am not alone with the hotspot struggles!

Danny